sites/valentin/requirements.txt
flask
pandas
scipy
cryptography
requests
numpy
python-Levenshtein
authlib
